+++ containers/threadsafe/lru_set.go
@@ -44,10 +44,10 @@ func (s *LruSet[T]) Insert(t T) {
 	}
 }
 
-func (s *LruSet[T]) NumItems() uint64 {
+func (s *LruSet[T]) NumItems() int {
 	s.RLock()
 	defer s.RUnlock()
-	return uint64(s.items.Len())
+	return s.items.Len()
 }
 
 func (s *LruSet[T]) Has(t T) bool {
